+++ packages/db/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @opengeni/db
 
+## 0.22.3
+
+### Patch Changes
+
+- Updated dependencies [e62495f]
+- Updated dependencies [b4982fa]
+- Updated dependencies [b4982fa]
+  - @opengeni/contracts@0.31.2
+  - @opengeni/config@0.10.2
+
 ## 0.22.2
 
 ### Patch Changes